requirements.txt
altair==4.1.0
annoy==1.17.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
bokeh==2.4.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cloudpickle==2.0.0
colorcet==3.0.0
cycler==0.11.0
dask==2021.11.2
datashader==0.13.0
datashape==0.5.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distributed==2021.11.2
entrypoints==0.3
fsspec==2021.11.0
gitdb==4.0.9
GitPython==3.1.24
graphviz==0.19.1
HeapDict==1.0.1
holoviews==1.14.6
idna==3.3
imageio==2.13.1
imbalanced-learn==0.9.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
llvmlite==0.37.0
locket==0.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
multipledispatch==0.6.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numba==0.54.1
numpy==1.20.3
packaging==21.2
pacmap==0.5.3
pandas==1.3.4
pandocfilters==1.5.0
panel==0.12.5
param==1.12.0
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==6.0.0
pycparser==2.21
pyct==0.4.8
pydeck==0.7.1
Pygments==2.10.0
pynndescent==0.5.5
pyparsing==2.4.7
pyrsistent==0.18.0
pysubgroup==0.7.2
python-dateutil==2.8.2
pyts==0.12.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.1.0
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
streamlit==1.1.0
tblib==1.7.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
umap-learn==0.5.2
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
xarray==0.20.1
zict==2.0.0
zipp==3.6.0
